+++ src/containers/Settings/Security/components/TOTP/TOTP.tsx
@@ -1,9 +1,11 @@
 import React, { FC, useState, useEffect } from 'react'
+import * as Yup from 'yup'
+import { useSnackbar } from 'notistack'
+import { useFormik } from 'formik'
 import { useMutation } from '@apollo/react-hooks'
-import { Dialog, DialogContent, DialogTitle } from '@material-ui/core'
-import { CREATE_RECOVERY_CODES, CREATE_TOTP } from '../../typeDefs'
-import QRCode from './QRCode'
-import RecoveryCodes from '../RecoveryCode/RecoveryCodes'
+import { Button, Card, Typography, TextField, Dialog } from '@material-ui/core'
+import classNames from 'classnames'
+import { CREATE_TOTP, VALIDATE_TOTP } from '../../typeDefs'
 import { goBack } from 'src/shared/utils'
 import styles from './totp.module.scss'
 
@@ -13,12 +15,45 @@ interface Props {
 
 const TOTP: FC<Props> = ({ showModal }) => {
   const userId = window.localStorage.getItem('userId')
-
   const [qrcode, setQRCode] = useState('')
-  const [recoveryCodes, setRecoveryCodes] = useState<string[]>([])
+  const { enqueueSnackbar } = useSnackbar()
+
+  const validationSchema = Yup.object().shape({
+    token: Yup.string()
+      .matches(/^\d{6}$/, 'token must be a six-digit code.')
+      .required('token should not be empty'),
+  })
+
+  const initialValues = {
+    token: '',
+  }
+
+  const {
+    handleSubmit,
+    getFieldProps,
+    resetForm,
+    isSubmitting,
+    errors,
+  } = useFormik({
+    initialValues,
+    validationSchema,
+    onSubmit: async values => {
+      await validateTOTP({
+        variables: { input: { ...values, userId } },
+      })
+      resetForm()
+    },
+  })
 
   const [createTOTP] = useMutation(CREATE_TOTP)
-  const [createRecoveryCodes] = useMutation(CREATE_RECOVERY_CODES)
+
+  const [validateTOTP] = useMutation(VALIDATE_TOTP, {
+    onCompleted() {
+      enqueueSnackbar('Two-factor authentication is available now!', {
+        variant: 'success',
+      })
+    },
+  })
 
   useEffect(() => {
     const fetchTOTPAndRecoveryCodes = async () => {
@@ -26,26 +61,50 @@ const TOTP: FC<Props> = ({ showModal }) => {
         variables: { userId },
       })
 
-      const recoveryCodesRequest = await createRecoveryCodes({
-        variables: { userId },
-      })
-
       setQRCode(TOTPRes.data.createTOTP.qrcode)
-      setRecoveryCodes(
-        recoveryCodesRequest.data.createRecoveryCodes.recoveryCodes,
-      )
     }
 
     fetchTOTPAndRecoveryCodes()
-  }, [createRecoveryCodes, createTOTP, userId])
+  }, [createTOTP, userId])
 
   return (
     <Dialog open={!!showModal} onClose={goBack}>
-      <DialogTitle>Recovery codes</DialogTitle>
-      <DialogContent>
-        {/* <RecoveryCodes recoveryCodes={recoveryCodes} /> */}
-        <QRCode userId={userId} qrcode={qrcode} />
-      </DialogContent>
+      <Typography variant="h5" gutterBottom>
+        Scan this barcode with your app.
+      </Typography>
+      <p className={styles.tips1}>
+        Scan the image above with the two-factor authentication app on your
+        phone.
+      </p>
+      <Card className={styles.qrcodeWrapper}>
+        <img src={qrcode} alt="qrcode" />
+      </Card>
+      <p className={classNames(styles.tipsBlod, styles.inputTipHeader)}>
+        Enter the six-digit code from the application
+      </p>
+      <p className={styles.tips1}>
+        After scanning the barcode image, the app will display a six-digit code
+        that you can enter below.
+      </p>
+      <form className={styles.customForm} onSubmit={handleSubmit}>
+        <TextField
+          error={!!errors.token}
+          helperText={errors.token}
+          autoFocus
+          {...getFieldProps('token')}
+        />
+
+        <Button
+          type="submit"
+          variant="contained"
+          color="primary"
+          size="small"
+          className={styles.inputBtn}
+          disabled={isSubmitting}
+        >
+          Enable
+        </Button>
+      </form>
     </Dialog>
   )
 }
